+++ extra_platforms/__init__.py
@@ -207,7 +207,7 @@ from .operations import (  # noqa: E402
     reduce,
 )
 
-__version__ = "3.2.0"
+__version__ = "3.2.1"
 """Examples of valid version strings according :pep:`440#version-scheme`:
 
 .. code-block:: python