+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [3.2.1 (unreleased)](https://github.com/kdeldycke/extra-platforms/compare/v3.2.0...main)
+
+> [!IMPORTANT]
+> This version is not released yet and is under active development.
+
 ## [3.2.0 (2025-05-16)](https://github.com/kdeldycke/extra-platforms/compare/v3.1.0...v3.2.0)
 
 - Add detection of CI systems: `GITHUB_CI`, `GITLAB_CI` and `UNKNOWN_CI`.